Optimize Hoppers with Plugin Heuristics
Use Heuristics on plugin behavior to detect potential intent of plugins using InventoryMoveItemEvent.

Mode 2:
If the plugin appears to only be checking the source/target inventory, IE to determine protection
concepts to cancel the event, then we can greatly reduce the number of event fires for IMIE
by only firing IMIE on the first event item attempted to be moved by a hopper.

Because the plugin doesn't access the item, then there is no reason to fire the event again
as the results should continue to be the same.

The only risk here is if the plugin was maintaining its own state and conditionally changed cancel behavior
based on that state, which is extremely unlikely.

Mode 3:
If no plugins are listening, we can go to an even more cleaner code path with no event fires.
This also will let server owners ignore plugins listening to it if they dont wish to let a
plugin ever cancel a hopper move.

Mode 0:
Standard CraftBukkit processing, however we have now made great optimizations to this code
so even mode 0 should be significantly faster than Spigot.

Mode 1: Auto detects if mode 2 or 3 is safe, else falls back to 0.